+++ src/pages/SignUp/SignUpStep3.vue
@@ -51,7 +51,7 @@
 <script setup>
 import { ref } from 'vue';
 import { useRouter } from 'vue-router';
-import { useRegisterStore } from '@/store/registerStore';
+import { useRegisterStore } from '@/stores/registerStore';
 import SignInUpFooter from '@/shared/components/SignInUpFooter.vue';
 import SignUpTitleStepper from '@/shared/components/SignUpTitleStepper.vue';
 import ButtonComponent from '@/shared/components/ButtonComponent.vue';
