+++ src/stores/exemplo.ts
